cleanup
git-svn-id: svn://svn.code.sf.net/p/jsidplay2/code/trunk@3112 05cac4c1-a8e4-46c3-9d4b-80028926d4ef

+++ jsidplay2/src/main/java/ui/menubar/MenuBar.java
@@ -259,10 +259,10 @@ public class MenuBar extends C64VBox implements UIPart {
 					node.setManaged(!newValue);
 					stage.setResizable(!newValue);
 					if (newValue) {
-						sidplay2Section.setMinimizedX((int) getScene().getWindow().getX());
-						sidplay2Section.setMinimizedY((int) getScene().getWindow().getY());
-						sidplay2Section.setMinimizedWidth((int) getScene().getWindow().getWidth());
-						sidplay2Section.setMinimizedHeight((int) getScene().getWindow().getHeight());
+						sidplay2Section.setMinimizedX(getScene().getWindow().getX());
+						sidplay2Section.setMinimizedY(getScene().getWindow().getY());
+						sidplay2Section.setMinimizedWidth(getScene().getWindow().getWidth());
+						sidplay2Section.setMinimizedHeight(getScene().getWindow().getHeight());
 						stage.sizeToScene();
 						Platform.runLater(() -> minimizeMaximize.setDisable(false));
 					} else {
